requirements.txt
# Core ML/AI libraries
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.24.3
pandas==2.0.3

# Text processing
nltk==3.8.1

# API framework (for future)
fastapi==0.104.1
uvicorn==0.24.0

# Development tools
pytest==7.4.0
python-dotenv==1.0.0

# Optional: LLM integration
google-generativeai==0.3.0  # For Gemini (optional)

# Utilities
tqdm==4.66.0
scikit-learn==1.3.0
